Consolidated extended sg client impl. (#371)
* consolidated ext sg client

* updated tool pg
