+++ packages/web/src/components/tools/ToolBuilder.tsx
@@ -7,7 +7,7 @@ import { useToast } from '@/src/hooks/use-toast';
 import { needsUIToTriggerDocFetch } from '@/src/lib/client-utils';
 import { formatBytes, generateUniqueKey, MAX_TOTAL_FILE_SIZE_TOOLS, processAndExtractFile, sanitizeFileName, type UploadedFileInfo } from '@/src/lib/file-utils';
 import { cn, composeUrl, getIntegrationIcon as getIntegrationIconName, getSimpleIcon, inputErrorStyles } from '@/src/lib/general-utils';
-import { Integration, IntegrationInput, SuperglueClient, Workflow as Tool, UpsertMode } from '@superglue/client';
+import { Integration, IntegrationInput, Workflow as Tool, UpsertMode } from '@superglue/client';
 import { integrationOptions, generateDefaultFromSchema } from "@superglue/shared";
 import { waitForIntegrationProcessing } from '@superglue/shared/utils';
 import { Validator } from 'jsonschema';
@@ -23,6 +23,7 @@ import { Switch } from '../ui/switch';
 import { Tabs, TabsList, TabsTrigger } from '../ui/tabs';
 import { Textarea } from '../ui/textarea';
 import JsonSchemaEditor from '../utils/JsonSchemaEditor';
+import { ExtendedSuperglueClient } from '@/src/lib/extended-superglue-client';
 
 type ToolBuilderView = 'integrations' | 'instructions';
 
@@ -81,33 +82,6 @@ const isMeaningfulResponseSchema = (schemaText: string | null): boolean => {
   }
 };
 
-class ExtendedSuperglueClient extends SuperglueClient {
-  async generateInstructions(integrations: IntegrationInput[]): Promise<string[]> {
-    let instructions: string[] = [];
-    const response = await fetch(`${this['endpoint']}/graphql`, {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-        'Authorization': `Bearer ${this['apiKey']}`
-      },
-      body: JSON.stringify({
-        query: `
-          query GenerateInstructions($integrations: [IntegrationInput!]!) {
-            generateInstructions(integrations: $integrations)
-          }
-        `,
-        variables: { integrations }
-      })
-    });
-    const result = await response.json();
-
-    instructions = result.data.generateInstructions;
-    if (instructions.length === 1 && instructions[0].startsWith('Error:')) {
-      throw new Error(instructions[0].replace('Error: ', ''));
-    }
-    return instructions;
-  }
-}
 
 export function ToolBuilder({
   initialView = 'integrations',
